pyproject.toml
[project]
name = "salesforce-async-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=47.0.0",
    "duckdb>=1.5.2",
    "faker>=40.15.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "msgspec>=0.21.1",
    "openpyxl>=3.1.5",
    "pyjwt>=2.12.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.27",
    "reportlab>=4.5.0",
    "rich>=15.0.0",
    "streamlit>=1.57.0",
    "structlog>=25.5.0",
    "textual>=8.2.5",
    "uvicorn>=0.46.0",
]
